feat(forum): save image to storage and add comment count

+++ cc/app/forum.py
@@ -1,25 +1,47 @@
 from flask import Blueprint, request, jsonify
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from config.firebase_config import db
+from firebase_admin import storage
 import logging
 from datetime import datetime
+import uuid
 
 forum_bp = Blueprint('forum_bp', __name__)
 
 @forum_bp.route('/forum', methods=['POST'])
 @jwt_required()
 def create_post():
-    data = request.get_json()
+    data = request.form
     user_id = get_jwt_identity()
     post_text = data.get('post_text')
-    post_image = str(data.get('post_image'))
+    post_image_file = request.files.get('post_image')
 
     try:
+        user_ref = db.collection('users').document(user_id)
+        user_doc = user_ref.get()
+        if not user_doc.exists:
+            raise Exception("User not found")
+
+        user_data = user_doc.to_dict()
+
+        if post_image_file:
+            image_filename = f"{uuid.uuid4()}-{post_image_file.filename}"
+            firebase_bucket = storage.bucket('cocodiag.appspot.com')
+            blob = firebase_bucket.blob(f"forums/{user_id}/{image_filename}")
+            post_image_file.seek(0)
+            blob.upload_from_file(post_image_file, content_type= post_image_file.content_type)
+            image_url = blob.public_url
+        else:
+            image_url = None
+
         doc_ref = db.collection('forum').document()
         doc_ref.set({
             "user_id": user_id,
+            "user_image": user_data.get('imageProfile', ''),
+            "user_name": user_data.get('name', ''),
+            "user_email": user_data.get('email', ''),
             "post_text": post_text,
-            "post_image": post_image,
+            "post_image": image_url,
             "count_like": 0,
             "count_comment": 0,
             "created_at": int(datetime.now().timestamp()),
@@ -31,6 +53,7 @@ def create_post():
 
         return jsonify(post_data), 201
     except Exception as e:
+        logging.error(f"Create post error: {e}")
         return jsonify({"message": str(e)}), 400
 
 @forum_bp.route('/forum', methods=['GET'])
@@ -71,8 +94,18 @@ def get_post(post_id):
 @forum_bp.route('/forum/<post_id>', methods=['DELETE'])
 @jwt_required()
 def delete_post(post_id):
+    user_id = get_jwt_identity()
     try:
         post_ref = db.collection('forum').document(post_id)
+        post_doc = post_ref.get()
+
+        if not post_doc.exists:
+            raise Exception("Post not found")
+
+        post_data = post_doc.to_dict()
+        if post_data["user_id"] != user_id:
+            return jsonify({"message": "You are not authorized to delete this post"}), 403
+
         post_ref.delete()
         return jsonify({"message": "post has been deleted"}), 200
     except Exception as e:
@@ -122,6 +155,14 @@ def create_comment():
         if not post_doc.exists:
             raise Exception("Post not found")
 
+        post_data = post_doc.to_dict()
+        if comment_text:
+            post_data["count_comment"] += 1
+        else:
+            raise Exception("Comment can't be empty")
+        
+        post_ref.update({"count_comment": post_data["count_comment"]})
+        
         comment_ref = post_ref.collection('comments').document()
         comment_ref.set({
             "user_id": user_id,
@@ -129,9 +170,12 @@ def create_comment():
             "created_at": int(datetime.now().timestamp())
         })
 
-        post_data = post_doc.to_dict()
         comments = post_ref.collection('comments').stream()
-        comments_list = [{"user_id": comment.get('user_id'), "comment_id": comment.id, "comment": comment.get('comment'), "created_at": comment.get('created_at')} for comment in comments]
+        comments_list = [{"user_id": comment.get('user_id'), 
+                          "comment_id": comment.id, 
+                          "comment": comment.get('comment'), 
+                          "created_at": comment.get('created_at')} 
+                          for comment in comments]
 
         return jsonify({"post_id": post_id, "comments": comments_list}), 201
     except Exception as e:
@@ -141,8 +185,24 @@ def create_comment():
 @forum_bp.route('/comment/<comment_id>', methods=['DELETE'])
 @jwt_required()
 def delete_comment(comment_id):
+    user_id = get_jwt_identity()
     try:
-        comment_ref = db.collection_group('comments').document(comment_id)
+        posts_ref = db.collection('forum').stream()
+        comment_ref = None
+        for post in posts_ref:
+            comment_doc_ref = db.collection('forum').document(post.id).collection('comments').document(comment_id)
+            comment_doc = comment_doc_ref.get()
+            if comment_doc.exists:
+                comment_ref = comment_doc_ref
+                break
+
+        if not comment_ref:
+            raise Exception("Comment not found")
+
+        comment_data = comment_doc.to_dict()
+        if comment_data["user_id"] != user_id:
+            return jsonify({"message": "You are not authorized to delete this comment"}), 403
+
         comment_ref.delete()
         return jsonify({"message": "comment has been deleted"}), 200
     except Exception as e:
